+++ include/simulation.hpp
@@ -33,7 +33,7 @@ private:
     std::unique_ptr<Communication> communication_;
 
     int maxCycle_;
-    double hyperPeriod_;
+    float hyperPeriod_;
 
     int numberOfTasks_;
     int numberOfRunnables_;
@@ -63,8 +63,8 @@ private:
         communication_->GetCommunicationTable(dag_, numberOfRunnables_, hyperPeriod_, runnablePermutation, runnableCommunications);
     }
 	
-    double GetReactionTime(std::map<std::pair<int, int>, std::vector<ExecutionInformation>>& processExecutions);
-    double GetDataAge(std::vector<int> executionPermutationPointer,
+    float GetReactionTime(std::map<std::pair<int, int>, std::vector<ExecutionInformation>>& processExecutions);
+    float GetDataAge(std::vector<int> executionPermutationPointer,
                       std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableExecutions,
                       std::map<std::pair<int, int>, std::vector<ExecutionInformation>>& processExecutions);
